PetScan: drop allow_nested attribute
The attribute was turned on by default with no way of turning it off.
The code that depended on allow_nested being off was therefore never nested.
Furthermore, we will be allowing for an encapsulation of dynamic
behavior in the near future, so there should be less need for turning
the allow_nested option off.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>