tree2scop.c: scop_from_non_affine_while: drop preallocation of statement number
Before the previous commit, we would extract a scop for the body
before extracting a scop for the non-affine condition.
Since we wanted the statement number of the statement evaluating
the non-affine condition to be smaller than the statement numbers
of the statements in the body, we would preallocate a statement number.
Now that we extract a scop from the body after extracting the scop
for the non-affine condition, this preallocation is no longer needed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>