pet_context: keep track of domain set rather than domain space
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
The constraints derived from the outer constructs will be stored
in the pet_context, which means that it needs to keep track of
a set rather than of just a space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>