pet_expr_filter: avoid introduction of constraints in index expression
For non-virtual accesses, the index expressions correspond
to the expressions that appear in the code without any additional
constraints on the domain that may appear in the corresponding
access relations.

Since filters currently start out as zero-dimensionsal, the constraints
on the domain naturally drop out since there is no place to attach
this information.  The only expections are affine filters since
they are one-dimensional instead of zero-dimensional.

However, when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, the filters
will be multi-dimensional from the start and we want to avoid
introducing domain constraints in their index expressions.
Removing these domain constraints currently only affects affine filters.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>