pet_create_test_index: allow specification of domain space
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
Whereas, currently, the index expression for virtual arrays
starts out as a mapping from an anonymous 0D space
that is modified as we move up the pet_tree,
in the top-down process, the statement domain space will be known
by the time this function is called and we will create an identity
mapping on this space instead.

We allow the users to pass this domain space to pet_create_test_index.
In the top-down process, the constraints derived from the outer
constructs will be stored in the pet_context.
We therefore obtain the domain space from the pet_context whenever
it is already available and take the current anonymous 0D space
as input otherwise.  The latter will be changed to take
the domain space from the pet_context in subsequent commits.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>