tree2scop.c: scop_from_{break,continue}: allow specification of domain space
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
Whereas, currently, the skip conditions of break and continue statements
are defined as mappings from an anonymous 0D space, this space
will include all outer loop iterators in the top-down process.

We allow the users to pass the domain space to scop_from_continue and
scop_from_break.
In the top-down process, the constraints derived from the outer
constructs will be stored in the pet_context.
We therefore obtain the domain space from the pet_context.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>