pet_scop constructors: allow specification of domain space
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
Currenlty, the context of a pet_scop always lives in a pure parameter
space, but this will be changed to the space that includes
all outer loop iterators during the top-down process.

We allow the users to pass the domain space to pet_scop_from_pet_stmt and
pet_scop_empty.
We therefore obtain the domain space from the pet_context whenever
it is already available and take the current anonymous 0D space
as input otherwise.  The latter will be changed to take
the domain space from the pet_context in subsequent commits.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>