pet_skip_info_*_extract*: pass along pet_context
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
In this top-down process, the pet_context will keep track of
the curent iteration domain, which is needed for the creation
of the statement evaluating the skip conditions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>