pet_scop_add_boolean_array: allow specification of domain constraints
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
Whereas, currently, the extent of virtual arrays
starts out as a universe 0D set that is modified as we move up the pet_tree,
in the top-down process, the domain constraints will be known
by the time this function is called and we will create an array
with an extent that matches the domain constraints from the start.

We allow the users to pass this domain to pet_scop_add_boolean_array.
In the top-down process, the constraints derived from the outer
constructs will be stored in the pet_context.
We therefore obtain the domain constraints from the pet_context.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>