tree2scop.c: handle_writes: separate out handling of assignments
handle_writes would recognize a top-level assignment in the given scop
as one with a universe domain.
However, when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, the domain of
the top-level assignment will no longer have a universe domain.

Separate out the handling of assignments into a separate function
that takes a pet_tree as input rather than a pet_scop.
These pet_trees always correspond to assignments that appear in
the source code, so there is no longer any need to check
for virtual arrays as they could not possibly have been assigned
in the source code.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>