tree2scop.c: valid_for_each_iteration: handle higher dimensional domains
Currently, the input sets of valid_for_each_iteration are always
single-dimensional.
However, when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, the dimension
of these sets will be equal to the number of outer loops,
including the current loop.
We therefore need to take into account that the dimension of
the input may be greater than 1.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>