pet_context_evaluate_expr: first insert domain
The pet_context currently has a zero-dimensional anonymous domain,
which is the same as the domains of the access expressions inside
the input pet_tree.
However, when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, the domain
of the pet_context will include references to outer constructs
and may therefore not be zero-dimensional.
We therefore need to map all those access expressions to the domain
of the pet_context before combining them with expressions
inside the pet_context.
This mapping currently has no effect, but will take effect as soon
as the domain of the pet_context changes.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>