scan.cc: clear_assignment: fix logic
The remainder of the method is supposed be executed when no value
was assigned to "decl" before, but we would actually return early
from the method when this condition holds.
This means that we would repeatedly perform the subsequent operations
even when they should no longer be needed, but, more importantly,
we could end up in the wrong state after the first call.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>