pet_expr_resolve_nested: allow specification of domain space
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
Currently, the domains of the access expresions inside the argument
of pet_expr_resolve_nested still refer to the 0D anonymous space.
After switching to a top-down process, the domains will refer
to the outer constructs.  We therefore need to take into account
that the domain may not by 0D.

We allow the users to pass the domain space to pet_expr_resolve_nested.
In the top-down process, the constraints derived from the outer
constructs will be stored in the pet_context.
We therefore obtain the domain space from the pet_context.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>