pet_extract_nested_from_space: embed arguments in given space
For the call from pet_expr_extract_nested, the embedding has no
effect since the given space is currently the same as the original space.
This will change in the next commit.
For the call from pet_stmt_extract_nested, this early embedding
replaces the embedding in pet_stmt_resolve_nested.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>