pet_extract_nested_from_space: add extra dimensions if needed during embedding
Currently, the domains of the expressions in nested parameters
are all zero-dimensional.
However, when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, the domains
of these expressions correspond to the outer constructs.
Since the expression may be used in a context with more outer constructs,
we may have to add extra dimensions during the embedding of
the expressions in the domain space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>