tree2scop.c: map_to_previous: handle higher dimensional domains
Currently, the input set of map_to_previous is always single-dimensional.
However, when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, the dimension
of this set will be equal to the number of outer loops,
including the current loop.
We therefore need to take into account that the dimension of
the input may be greater than 1.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>