separate out pet_context_evaluate_expr
Evaluating an expression in the context of a pet_context
currently amounts to plugging in the arguments of the access expressions,
but as we move closer to our goal of turning the extraction of a pet_scop
from a pet_tree into a top-down process, this will be extended
with a conversion of scalar reads and the introduction of an iteration domain.

With this commit, we also make a clearer distinction between expressions
that have not been "evaluated" yet, i.e., those that come straight
from the input pet_tree and those that have been evaluated already.
While plugging in arguments is essentially idempotent,
the introduction of an iteration domain that we will need to do
can only be performed once.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>