pet_scop_restrict: also restrict virtual array extents
When we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, these extents
will end up getting restricted automatically.  We perform
the restriction in advance such that the test cases are not
changed by the commit that performs the actual change from
bottom-up to top-down.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>