PetScan::extract_for: do not allow initial value to depend on iterator
If the initializer of a for loop refers to the loop iterator,
then it refers to the value before the loop, meaning that
the initial value is not an affine expression.

No test case is added here since a test case that tickles
this bug would result in an error message after the fix.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>