tree2scop.c: scop_from_block: collect affine continue conditions
In the previous commit, we would only add the latest affine skip
conditions derived from continue statements.  We did this mainly
to avoid any changes in the test cases.
We now switch to collecting all such conditions such that we
can exploit any previous such conditions as well.
This has an effect on the continue5 test case, which now produces
a __pet_test_1 virtual array that is identical to the __pet_test_0
virtual array.  Although this is somewhat suboptimal, it is no
worse than the situation before this commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>