add functions for manipulating the domain of a pet_context
A pet_context currently always has a zero-dimensional domain,
but when we change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process, we will want
to keep track of the outer loop constraints in the pet_context.
These functions will allow us to do so.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>