PetScan::extract_scop: also add all pet_arrays
We now call scan_arrays in extract_scop rather than in scan to ensure that
scan_arrays is executed in the same pet_context.
In particular, in subsequent commits, we will be modifying the pet_context
inside extract_scop and we want these modifications to be available to
scan_arrays as well.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>