add pet_context_get_gist_domain
Now that we construct a pet_scop from the top down, we have the domain
information from outer loop constructs available while extracting from
subtrees.  We can therefore exploit the information to simplify
expressions as soon as they are created.
We have to be careful though not to introduce any nested expression
parameters during such a simplification.  We therefore add
a pet_context_get_gist_domain function that return the domain
with the nested expression parameters removed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>