scan.cc: indicator_function: coalesce domain before using it in intersection
The domain passed to indicator_function is obtained from isl_pw_aff_domain
or an intersection of the results of two such calls.
Since isl_pw_aff_domain currently simply collects the pieces, this domain
may consist of several pieces and there is no need to cut
the result of indicator_function along the same pieces.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>