PetScan::extract_expr(CallExpr *): extract function summary from function body
We want to derive the accesses performed through a function call
from the accesses in the body of the function that is being called.
In order to do so, we abstract the accesses of the function
in a function summary.  In a later commit we will instantiate
the function summary at the location of the function call.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>