derive access relations of call expressions from function summary
In previous commits, we have added the infrastructure to store and
extract function summaries from function bodies.  In this commit,
we exploit this information to derive access relations
from the function summaries whenever they are available.
This allows for an extraction of more accurate access relations than
those based on the default heuristic that the entire array slice
passed to the function is either read or written, which may not
only be wildly inaccurate, but it may also be wrong.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>