pet_scop_from_pet_stmt: do not extract context from kill statements
The argument of a kill statement is the extent of the killed array.
Extracting context constraints from this extent restricts the context
to those parameter values where the array is non-empty.
If we want to restrict the context in this way, we should do
so explicitly and not by acciddent through the argument of a kill statement.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>