tree2scop.c: extract_kill: update kill expression rather than creating new one
That is, take the kill expression from the other statement, but
drop the statement name from the expression, so that we can add
the name of the new statement.

As a nice side-effect of the clean-up, we remove a call to
pet_expr_access_get_access, which we want to remove soon.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>