nest.c: move is_sub_access to expr.c
This function is intimately tied to both its use by mark_self_dependences
and the internal representation of pet_expr and could therefore belong
in either nest.c or expr.c.  Since we want to remove
pet_expr_access_get_access, it becomes more interesting to move it
to expr.c.  The function can then also be adjusted to the fact
that the access relation may not always be available.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>