drop pet_expr_access_get_access
In a future commit, we will only keep track of access relation(s)
when they cannot be trivially derived from the index expression.
In principle it would be possible to keep pet_expr_access_get_access and
to generate an access relation on-the-fly, but it seems best to ensure
that outside users of pet_expr objects do not make use of any access
relation as we will also change the type of the access relations
in a future commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>