pet_expr_tag_access: avoid use of target space
In the next commit, we will change the functions that return
access relations to differentiate between may_read, may_write
and must_write and to return union maps instead of maps.
This means that pet_expr_tag_access will return an isl_union_map and
therefore also take an isl_union_map as input.
We can then no longer refer to the target space of the access relation
as there may be several target spaces.  There is however no need
to refer to this target space.  Perform the tagging directly based
on the domain space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>