PetScan::extract_argument: handle passing of entire arrays
PetScan::extract_argument uses a simple heuristic to determine
whether a slice of an array passed to a function should be considered
a read or a write of the slice.  (In particular, if the function
is declared as taking a pointer to a const type, then it is considered
a read, otherwise it is considered a write).
However, it would only consider proper slices of an array and not
entire arrays.  Consider them too for consistency.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>