explicitly mark kill accesses
The access expression argument of a kill operation represents
the set of array elements that are killed by the operation.
This access expression is marked neither read nor write.
Explicitly mark it as a kill access for clarity.

The explicit marking will also be useful in printing different
labels for the access relation.  It is simply called "relation" now,
but when we split the relation into may_read, may_write and must_write,
then none of these labels are suitable for a kill access.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>