introduce pet_function_summary
We want to derive the accesses performed through a function call
from the accesses in the body of the function that is being called.
In order to do so, we will abstract the accesses of the function
in a function summary and then instantiate the function summary
at the location of the function call.

This commit introduces the abstraction that keeps track
of the function summary.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>