Makefile.am: add dummy C++ source file to pet_scop_cmp and pet_codegen
This ensures that pet_scop_cmp and pet_codegen will be linked using the
C++ compiler rather than the C compiler, which appears to be required
when libpet is constructed as a static library.

This commit complements the changes in 9f7f5a6 (Makefile.am: add dummy
C++ source file to pet and pet_check_code, Tue Sep 3 13:40:10 2013 +0200).
Back then, the problem only appeared to occur for pet and pet_check_code,
but depending on the phase of the moon, the problem now also appears
to occur for pet_scop_cmp.  Handle pet_codegen as well to ensure
we will not run into this issue again later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>