pet_expr: keep track of function summaries in call expressions
We want to derive the accesses performed through a function call
from the accesses in the body of the function that is being called.
In order to do so, we will abstract the accesses of the function
in a function summary and then instantiate the function summary
at the location of the function call.

This commit allows a call expression to keep track of the summary
of the called function.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>