pet_expr_access_set_access: set read/write on non-empty access relation
For consistency, we should mark an access expression as reading/writing
as soon as one of its reading/writing access relations is non-empty.

This change does not have any effect at the moment, but when we start
handling function summaries, it will be convenient to have the read/write
attributes automatically updated when we set one of the reading or writing
access relations to a non-empty relation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>