tree2scop.c: is_nested_allowed: do not allow nesting if current level has break
The original code would check for a "now" skip.  In 417afba (tree2scop.c:
is_nested_allowed: take tree instead of scop, Wed Jan 1 14:25:21 2014 +0100),
this was replaced by a check for a continue statement inside the tree.
However, if there is a break statement then we should also not allow
nesting.  A break statement will also result in the "now" skip being
set so the original code got it right.

Reported-by: Alexey Kravets <Alexey.Kravets@arm.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>