pet_expr_new_unary: take type size argument
We already keep track of the type size of binary operators,
but we did not do this for unary ones.
The main reason for keeping track of the type size of mainly
the unary minus operator is that, when applied to a constant
integer expression, it would be converted to an affine expression in
call_plug_in_affine_args in context.c so that it can be plugged
into the access relations of the array arguments.

The side effect is that all such function arguments are now
converted into access expression, whereas it may be nicer
to convert them to integer expressions.
We can handle that later if needed.

Reported-by: Riyadh BAGHDADI <baghdadi.m.riyadh@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>