PetScan::extract_argument: only perform write detection optionally
When we add support for __pencil_kill, this method will be called
on the arguments of the __pencil_kill, which has no prototype.
This call would therefore fail while checking whether the access
is a write or not.  This detection is irrelevant for __pencil_kill
arguments, so we allow the detection to be turned off.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>