pet_stmt_build_ast_exprs: ignore expression arguments not appearing in index
We would fail to build an AST expression for an access with arguments
that are not themselves accesses, even if these arguments are not needed
to build the AST expression because they are not referenced in the index
expression of the access.
Ignore those arguments instead and update pet_stmt_can_build_ast_exprs
accordingly.

Reported-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>