encapsulate loop containing continue/break when encapsulating dynamic control
If we decide to encapsulate a statement that contains a break
or a continue statement, then we need to make sure that we also
encapsulate the loop containing the break or continue.
Otherwise, we would not take into account the effect of
the break or continue on subsequent statements inside that loop.

Reported-by: Ali Zaidi <Ali.Zaidi@arm.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>