tree2scop.c: scop_from_affine_if: check condition of conditional assignment
Although the condition of the special case of a conditional assignment
is allowed not to be static-affine, the condition should still convert
to an affine expression when nested accesses are allowed in the expression.
We used to test for this condition, but it got dropped by mistake
in 2422264 (construct pet_scop from intermediate pet_tree data structure,
Tue Dec 31 12:00:45 2013 +0100).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>