skip empty scops when autodetect is turned on
The method PetScan::scan currently has no way of indicating
that no scop was found when autodetect is turned on.
It will therefore return an empty scop if no actual scop is found.
This scop cannot be used since it does not have any proper line
information.  It should therefore not be passed to the user.

Reported-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>