PetScan::extract_argument: mark non-const pointer arguments as potential writes
The initial version of pet followed the convention that if a function
takes a pointer to const that then the passed array is read,
while if it takes a pointer to non-const that then the passed array
is completely written.  This convention has remained even after
the introduction of the distinction between may-writes and must-writes.
Change it now to only mark the accessed elements as possibly being written to.
At the same time, retain the fact that the elements may also be read by
the function.  These access relations can always be refined by the user
by providing a summary function.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>