try and include initial declarations in outer block when autodetecting
When autodetection is turned on, we skip declarations in the outer
block to avoid extracting a scop that consists of only declarations.
However, once we have been able to extract at least one statement,
there is no harm in trying to include some of these initial declarations
in the scop.  Their inclusion is useful as it allows the declared
variables to be killed at the end of the outer block.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>