support inlining of outermost call expressions
In particular, any expression statement of which the outer expression
is a call to a function, the definition of which is marked "inline",
is replaced by an inlined copy of the called function.
Calls that appear as subexpressions are not inlined because
that would require support for return statements.

The inlining is performed at the level of (unevaluated) pet_trees
to ensure a consistent view of what variables are considered parameters.

The inlining is performed by assigning the scalar arguments and
the indices of the array arguments to temporary variables and
replacing the corresponding formal arguments in the tree extracted
from the function body by expression referring to or containing
these temporary variables.  The regular evaluation mechanism
is then used to propagate the scalar values whenever feasible.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>