pet_patch_union_map: make warning optional
pet_patch_union_map is currently only called on a pre-evaluated access relation.
If an initial index in this access relation needs to be zero, then it makes
sense to warn about cases where this initial index is not obviously zero.
In subsequent commits, pet_patch_union_map will also get called during
the transformation of unevaluated pet_trees.  In such cases, it may not
be clear yet that an initial index will definitely be zero and no warning
should be printed.  Make the printing of a warning optional such that
the caller can decide whether or not to print a warning.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>