add pet_substituter
A pet_substituter keeps track of a number of identifiers that need
to be replaced by an (address of an) access expression and applies
the substitutions to a pet_tree.
This will be useful for renaming, where an old name is substituted
by an access to the new name, and later on also for inlining.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>