if the same name is used for several variables, then rename at least one of them
The same name may be used for several variables in the program fragment
from which the scop is extracted.  The lexical scopes of these
variables are not reflected in the extracted scop and at least
one of these variables therefore needs to be renamed.

Note that renaming is only applied to variables that are declared
inside a compound statement that lies entirely inside the input
fragment.  That is, these variables are entirely local to the scop.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>