tree2scop.c: scop_from_tree_macro: drop redundant argument and plug leak
The redundant argument has been there from the start in 892c4f3
(add encapsulate_dynamic_control option, Tue Jan 28 11:16:24 2014 +0100).
Usually, the argument is NULL, so it does not have any effect,
but if the "macro-statement" has a label, then a copy would be passed
that is never freed, resulting in a memory leak.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>