rename pet_scop_collect_* to pet_scop_get_*
Although this functions effectively collect the information,
this is not relevant to the user of these functions.
From the point of view of the user, the main point of these
functions is that they return the specified information.

Additionally, rename pet_scop_collect_domains to pet_scop_get_instance_set.
The term "domain" is a bit too generic.  The term "instance set"
will be used by the "Presburger Formulas and Polyhedral Compilation" tutorial.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>