add isl_ctx_alloc_with_pet_options
The isl_ctx created by isl_ctx_alloc does not reference the pet options and
can therefore not be used in any pet operations.  Instead, the user needs
to construct an isl_ctx using isl_ctx_alloc_with_options with an option
hierarchy that includes the pet options.
The isl_ctx_alloc_with_pet_options function simplifies this process
in the case no extra options are required.
In particular, this will be useful for the upcoming python interface to pet.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>