pet_scop_collect_arrays: keep track of isl_id representing array
pet_scop_collect_arrays collects arrays and sub-arrays that have
not already been added to the pet_scop.  It would represent
the outer arrays by means of their variable declarations,
but these declarations do not take into account any possible renamings
that may have occurred to the variables.
This could result in array entries with the wrong names or even
missing array entries in case of conflicting names.

The isl_id from which the variable declaration is extracted,
does keep track of the new name, so use those directly
to represent the outer arrays.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>